requirements.txt
python-dotenv
contentful
contentful_management
flask
requests
